examen

+++ EntornoServidor/Examen12Dic2023MaldonadoPuertoManuel/joyitas_andalucia_alumnos/app/models/artistas.php
@@ -1,10 +1,41 @@
 <?php
-class ArtistasModel{
+
+ini_set('display_errors', 1);
+ini_set('display_startup_errors', 1);
+error_reporting(E_ALL);
+
+class ArtistasModel
+{
     private $dbh;
-    public function __construct(){
-        //Instanciar la conexión que estará disponible en el modelo.
-        $this->dbh = new PDO('mysql:host=localhost;dbname=joyitas_andalucia', 'joyitas_andalucia', 'joyitas');
+    public function __construct()
+    {
+        try {
+            //Instanciar la conexión que estará disponible en el modelo.
+            $this->dbh = new PDO('mysql:host=localhost;dbname=joyitas_andalucia', 'joyitas_andalucia', 'joyitas');
+        } catch (Exception $e) {
+            $error = error_get_last();
+
+
+            http_response_code(500);
+            if(http_response_code(500) == 500){
+                header('Location: ../app/views/error500.html');
+            }
+            
+            echo '<pre>';
+            echo '<br>';
+            echo 'Error 500: Internal Server Error';
+            echo '</pre>';
+
+            error_log('No se pudo conectar a la base de datos: ' . $e->getMessage());
+            
+            // Capturar el error si no hay conexion
+            die('No se pudo conectar a la base de datos. Por favor, inténtelo de nuevo más tarde.');
+        }
     }
+
+
+
+    
     public function getListArtistas($idArtista = [])
     {
         try {
@@ -50,43 +81,51 @@ class ArtistasModel{
     }
 
     public function borrarArtista($idArtista)
-{
+    {
 
-    try {
-        $this->dbh->beginTransaction();
-        //Eliminar conciertos del artista
-        $query = 'DELETE FROM artistas where id = :id';
+        try {
+            $this->dbh->beginTransaction();
+            //Eliminar conciertos del artista
+            $queryConciertos = 'DELETE FROM conciertos where id_artista = :id';
 
-        $stmt = $this->dbh->prepare($query);
-        $stmt->bindParam(':id', $idArtista, PDO::PARAM_INT);
+            $stmt = $this->dbh->prepare($queryConciertos);
+            $stmt->bindParam(':id', $idArtista);
 
-        if (!$stmt->execute()) {
-            throw new Exception("Error al ejecutar la eliminacion del artista: " . implode(" ", $stmt->errorInfo()));
+            if ($stmt->execute()) {
+                $query = 'DELETE FROM artistas where id = :id';
+
+                $stmt = $this->dbh->prepare($query);
+                $stmt->bindParam(':id', $idArtista, PDO::PARAM_INT);
+
+                if (!$stmt->execute()) {
+                    throw new Exception("Error al ejecutar la eliminacion del artista: " . implode(" ", $stmt->errorInfo()));
+                }
+            }
+            $this->dbh->commit();
+        } catch (Exception $e) {
+            $this->dbh->rollBack();
+            throw new Exception('Error al eliminar el artista: ' . $e->getMessage());
         }
-        $this->dbh->commit();
-    } catch (Exception $e) {
-        $this->dbh->rollBack();
-        throw new Exception('Error al eliminar el artista: ' . $e->getMessage());
     }
-}
 
-public function guardarArtista($nombre, $genero, $fechaNac, $precioBolo, $localNac)
-{
-    // INSERT INTO `cantantes` (`id`, `nombre`, `genero`, `fecha_nacimiento`, `precio_bolo`, `localidad_nacimiento`) VALUES (NULL, 'Manue', 'Rap', '2013-03-29', '43', 'Sevilla');
-    $query = 'INSERT INTO artistas (id, nombre, genero, fecha_nacimiento, precio_bolo, localidad_nacimiento) 
+
+    public function guardarArtista($nombre, $genero, $fechaNac, $precioBolo, $localNac)
+    {
+        // INSERT INTO `cantantes` (`id`, `nombre`, `genero`, `fecha_nacimiento`, `precio_bolo`, `localidad_nacimiento`) VALUES (NULL, 'Manue', 'Rap', '2013-03-29', '43', 'Sevilla');
+        $query = 'INSERT INTO artistas (id, nombre, genero, fecha_nacimiento, precio_bolo, localidad_nacimiento) 
     VALUES (null,:nombre, :genero, :fecha_nacimiento, :precio_bolo, :localidad_nacimiento)';
 
-    $stmt = $this->dbh->prepare($query);
-    $stmt->bindParam(':nombre', $nombre);
-    $stmt->bindParam(':genero', $genero);
-    $stmt->bindParam(':fecha_nacimiento', $fechaNac);
-    $stmt->bindParam(':precio_bolo', $precioBolo);
-    $stmt->bindParam(':localidad_nacimiento', $localNac);
+        $stmt = $this->dbh->prepare($query);
+        $stmt->bindParam(':nombre', $nombre);
+        $stmt->bindParam(':genero', $genero);
+        $stmt->bindParam(':fecha_nacimiento', $fechaNac);
+        $stmt->bindParam(':precio_bolo', $precioBolo);
+        $stmt->bindParam(':localidad_nacimiento', $localNac);
 
-    // Intentamos ejecutar la consulta
-    if (!$stmt->execute()) {
-        // Si hay un error, lanzamos una excepción con el mensaje de error
-        throw new Exception('Error al guardar el artista: ' . implode(" ", $stmt->errorInfo()));
+        // Intentamos ejecutar la consulta
+        if (!$stmt->execute()) {
+            // Si hay un error, lanzamos una excepción con el mensaje de error
+            throw new Exception('Error al guardar el artista: ' . implode(" ", $stmt->errorInfo()));
+        }
     }
-}
 }
